Fixup circular import for make_c_string
